pmx도 터프하게 렌더링

+++ PoseViewer/MmdPlayer.h
@@ -11,7 +11,7 @@ public:
 	virtual const DirectX::XMFLOAT3* GetVertices() const = 0;
 	virtual int GetVerticesCount() const = 0;
 
-	virtual const uint16_t* GetIndices() const = 0;
+	virtual const int* GetIndices() const = 0;
 	virtual int GetIndicesCount() const = 0;
 
 	virtual const DirectX::XMMATRIX* GetWorldTransform() const = 0;
